ocr-worker/requirements.txt
# --- Core Worker ---
pika==1.3.2
minio==7.2.5
psycopg2-binary==2.9.9
SQLAlchemy==2.0.29
python-dotenv==1.0.1
pillow>=10.2.0
structlog>=24.1.0

# --- Telemetry ---
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0

# --- AI & Resiliency (NEW) ---
anthropic>=0.25.0
pydantic>=2.6.4
tenacity>=8.2.3
redis>=5.0.1